rule_generator/requirements.txt
flask
openai
python-dotenv
pyyaml
requests
rapidfuzz